Moved out utils
